+++ app/Events/EventFinalInformation.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Events;
+
+use App\Models\Booking;
+use Illuminate\Broadcasting\Channel;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Broadcasting\PresenceChannel;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class EventFinalInformation
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    public $booking;
+    public $testMode;
+
+    /**
+     * Create a new event instance.
+     *
+     * @return void
+     */
+    public function __construct(Booking $booking, $testMode = false)
+    {
+        $this->booking = $booking;
+        $this->testMode = $testMode;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return \Illuminate\Broadcasting\Channel|array
+     */
+    public function broadcastOn()
+    {
+        return new PrivateChannel('channel-name');
+    }
+}