+++ app/Providers/EventServiceProvider.php
@@ -13,9 +13,24 @@ class EventServiceProvider extends ServiceProvider
      * @var array
      */
     protected $listen = [
-        'App\Events\Event' => [
-            'App\Listeners\EventListener',
+        'App\Events\BookingConfirmed' => [
+            'App\Listeners\SendBookingConfirmedNotification',
         ],
+        'App\Events\BookingCancelled' => [
+            'App\Listeners\SendBookingCancelledNotification',
+        ],
+        'App\Events\BookingChanged' => [
+            'App\Listeners\SendBookingChangedNotification',
+        ],
+        'App\Events\BookingDeleted' => [
+            'App\Listeners\SendBookingDeletedNotification',
+        ],
+        'App\Events\EventBulkEmail' => [
+            'App\Listeners\SendEventBulkEmailNotification',
+        ],
+        'App\Events\EventFinalInformation' => [
+            'App\Listeners\SendEventFinalInformationNotification',
+        ]
     ];
 
     /**